Issue #49: Rename package of feature-session to mozilla.components.feature.session.
