+++ components/feature/session/src/main/java/mozilla/components/feature/session/SessionFeature.kt
@@ -2,19 +2,22 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.toolbar
+package mozilla.components.feature.session
 
-import mozilla.components.toolbar.Toolbar
+import mozilla.components.engine.Engine
+import mozilla.components.engine.EngineView
 import mozilla.components.session.SessionManager
 
 /**
- * Feature implementation for connecting a toolbar implementation with the session module.
+ * Feature implementation for connecting the engine module with the session module.
  */
-class ToolbarFeature(
+class SessionFeature(
     sessionManager: SessionManager,
-    toolbar: Toolbar
+    engine: Engine,
+    engineView: EngineView,
+    sessionMapping: SessionMapping = SessionMapping()
 ) {
-    private val presenter = ToolbarPresenter(sessionManager, toolbar)
+    private val presenter = EngineViewPresenter(sessionManager, engine, engineView, sessionMapping)
 
     /**
      * Start feature: App is in the foreground.
