+++ components/feature/session/src/main/java/mozilla/components/feature/session/SessionProxy.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.toolbar
+package mozilla.components.feature.session
 
 import mozilla.components.engine.EngineSession
 import mozilla.components.session.Session
